增加背包,精英怪,掉落宝石,装备升级,宝石升级

+++ Assets/scripts/Server/ServerHandler/ShopHandler.cs
@@ -32,7 +32,7 @@ namespace ServerPacketHandler {
         public override void HandlePacket(KBEngine.Packet packet)
         {
             var inpb = packet.protoBody as ChuMeng.CGUseUserProps;
-            var ret = playerData.ReduceItem((int)inpb.UserPropsId);
+            var ret = playerData.ReduceItem(inpb.UserPropsId);
             var pb = ChuMeng.GCUseUserProps.CreateBuilder();
             if(ret) {
                 ChuMeng.ServerBundle.SendImmediate(pb, packet.flowId);
